backend/ocr_project/requirements.txt
Django>=4.2
djangorestframework
gunicorn
python-dotenv
mistralai
reportlab
python-docx

